# docs/openapi.yml
openapi: 3.1.0
info:
  title: Teman Berbahasa Management System API
  description: |
    Internal backend API for the Teman Berbahasa Tutor Place Management System.
    Staff and teacher-facing; no student-facing portal.

    ## Authentication
    All protected endpoints require a JWT Bearer token in the `Authorization` header.
    Tokens are RS256-signed. Access tokens expire in **15 minutes**; refresh tokens in **7 days**.

    ## Error Envelope
    All error responses use a consistent envelope:
    ```json
    { "error": { "code": "ERROR_CODE", "message": "Human-readable message" } }
    ```
    Validation errors additionally include a `fields` map:
    ```json
    { "error": { "code": "VALIDATION_ERROR", "message": "validation failed", "fields": { "email": "required" } } }
    ```

    ## Conventions
    - All IDs are UUID v7 strings.
    - All datetimes are ISO 8601 UTC.
    - Field names are `snake_case` throughout.
    - String fields are trimmed of leading/trailing whitespace before processing.
  version: 0.1.0
  contact:
    name: Hernando
    email: borosi.hernando@gmail.com
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.temanberbahasa.id
    description: Production (via Caddy reverse proxy, auto TLS)
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Log in
      description: |
        Authenticates a staff/teacher account with email and password.
        Returns an access token, a refresh token, and basic user info.

        **Account lock:** After 10 consecutive failed attempts the account is locked and
        returns `401 ACCOUNT_LOCKED`. Successful login resets the failure counter.

        **Inactive accounts** are rejected with `401 UNAUTHORIZED` (same response as bad
        credentials — no user enumeration).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              email: staff@temanberbahasa.id
              password: s3cur3P@ss
      responses:
        "200":
          description: Login successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
              example:
                access_token: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...
                refresh_token: a3f1d2c4e5b6...
                expires_in: 900
                user:
                  id: 019687a2-0000-7000-8000-000000000001
                  first_name: Siti
                  last_name: Rahayu
                  role: staff
        "401":
          description: Invalid credentials or account locked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid_credentials:
                  summary: Wrong email or password
                  value:
                    error:
                      code: UNAUTHORIZED
                      message: invalid email or password
                account_locked:
                  summary: Too many failed attempts
                  value:
                    error:
                      code: ACCOUNT_LOCKED
                      message: account is locked due to too many failed login attempts
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshToken
      summary: Refresh access token
      description: |
        Exchanges a valid refresh token for a new access token and a new refresh token
        (token rotation — the old refresh token is invalidated immediately).

        Reusing an already-rotated refresh token returns `401 UNAUTHORIZED`.
        Expired refresh tokens also return `401 UNAUTHORIZED`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
            example:
              refresh_token: a3f1d2c4e5b6...
      responses:
        "200":
          description: Token refreshed successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
              example:
                access_token: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...
                refresh_token: b9e2a1f3d4c5...
                expires_in: 900
        "401":
          description: Refresh token is invalid, expired, or already rotated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: UNAUTHORIZED
                  message: invalid or expired refresh token
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: Log out
      description: |
        Revokes the supplied refresh token and all other refresh tokens belonging to
        the same user (full device logout). Idempotent — submitting an unknown or
        already-revoked token still returns `204`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequest"
            example:
              refresh_token: a3f1d2c4e5b6...
      responses:
        "204":
          description: Logged out successfully (no body).
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /auth/forgot-password:
    post:
      tags: [Auth]
      operationId: forgotPassword
      summary: Request password reset
      description: |
        Sends a password-reset link to the supplied email address if it matches an
        active account. Always returns `200` with the same body regardless of whether
        the email exists — prevents user enumeration.

        The reset token embedded in the link expires after **1 hour**.
        Any previous unused reset tokens for the account are invalidated before the new
        one is issued.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
            example:
              email: staff@temanberbahasa.id
      responses:
        "200":
          description: Request processed (email may or may not have been sent).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
              example:
                message: "If that email exists, a reset link has been sent."
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /auth/reset-password:
    post:
      tags: [Auth]
      operationId: resetPassword
      summary: Reset password
      description: |
        Sets a new password using a one-time reset token obtained from the
        forgot-password email link.

        On success, all existing refresh tokens for the account are revoked so all
        active sessions are terminated.

        **Password requirements:** minimum 8 characters.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
            example:
              token: a3f1d2c4e5b6c7d8e9f0a1b2c3d4e5f6a7b8c9d0e1f2a3b4c5d6e7f8a9b0c1d2
              new_password: newS3cur3P@ss
      responses:
        "200":
          description: Password updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
              example:
                message: "Password updated. Please log in again."
        "400":
          description: Reset token is invalid or has expired.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: BAD_REQUEST
                  message: invalid or expired reset token
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /auth/me:
    get:
      tags: [Auth]
      operationId: getMe
      summary: Get current session
      description: |
        Returns the authenticated user's profile. Use to verify a session is still
        valid and retrieve the same user object returned by `POST /auth/login`.

        Requires a valid Bearer access token.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Session is valid; user info returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
              example:
                id: 019687a2-0000-7000-8000-000000000001
                first_name: Siti
                last_name: Rahayu
                role: admin
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256-signed JWT. Obtain via `POST /auth/login`. Expires in 15 minutes.
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          description: Registered staff/teacher email address. Trimmed before lookup.
        password:
          type: string
          description: Account password.
          minLength: 1
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          description: Opaque refresh token returned by login or a previous refresh call.
    LogoutRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          description: Opaque refresh token to revoke. All tokens for the user are revoked.
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          description: Email address associated with the account. Trimmed before lookup.
    ResetPasswordRequest:
      type: object
      required: [token, new_password]
      properties:
        token:
          type: string
          description: One-time reset token from the password-reset email link (hex-encoded, 64 chars).
        new_password:
          type: string
          minLength: 8
          description: New password. Minimum 8 characters.

    # ── Responses ────────────────────────────────────────────────────────────
...
    LoginResponse:
      type: object
      required: [access_token, refresh_token, expires_in, user]
      properties:
        access_token:
          type: string
          description: RS256-signed JWT. Valid for `expires_in` seconds.
        refresh_token:
          type: string
          description: Opaque token. Valid for 7 days. Use `POST /auth/refresh` to rotate.
        expires_in:
          type: integer
          description: Access token lifetime in seconds (always 900).
          example: 900
        user:
          $ref: "#/components/schemas/UserInfo"
    TokenPair:
      type: object
      required: [access_token, refresh_token, expires_in]
      properties:
        access_token:
          type: string
          description: New RS256-signed JWT. Valid for `expires_in` seconds.
        refresh_token:
          type: string
          description: New opaque refresh token. The previous token is now invalid.
        expires_in:
          type: integer
          description: Access token lifetime in seconds (always 900).
          example: 900
    UserInfo:
      type: object
      required: [id, first_name, last_name, role]
      properties:
        id:
          type: string
          format: uuid
          description: UUID v7 identifier of the authenticated user.
          example: 019687a2-0000-7000-8000-000000000001
        first_name:
          type: string
          example: Siti
        last_name:
          type: string
          example: Rahayu
        role:
          type: string
          enum: [admin, teacher, staff]
          description: User role. Controls access to protected endpoints.
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: Human-readable status message.
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code.
              enum:
                - BAD_REQUEST
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - GONE
                - VALIDATION_ERROR
                - RATE_LIMITED
                - ACCOUNT_LOCKED
                - INTERNAL_ERROR
            message:
              type: string
              description: Human-readable explanation.
            fields:
              type: object
              description: |
                Present only on `VALIDATION_ERROR`. Maps field name to the failing
                validation tag (e.g. `"email": "required"`, `"new_password": "min"`).
              additionalProperties:
                type: string
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: UNAUTHORIZED
              message: missing or invalid token
...
    ValidationError:
      description: Request body failed validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: VALIDATION_ERROR
              message: validation failed
              fields:
                email: required
                password: required
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: INTERNAL_ERROR
              message: internal server error